Move stacking functions out of utils into stacking.py
