+++ external/fv3fit/fv3fit/_shared/__init__.py
@@ -13,5 +13,6 @@ from .scaler import (
     NormalizeTransform,
 )
 from .predictor import Predictor
-from .utils import parse_data_path, StackedBatches, stack_non_vertical
+from .utils import parse_data_path
+from .stacking import StackedBatches, stack_non_vertical
 from .models import EnsembleModel, DerivedModel
